Separate exported modules and internal code
Reorganize the `src/` directory following [1], creating a clear
separation between code that is meant for exporting (modules) and only
for internal use (internal). This is purely refactoring, no functional
change should have occured.

Besides making the `src/` directory hopefully clearer, it also
simplifies configuration files, and especially improves clarity on what
tests should do what - unit tests test what is internal, other tests
test everything and thus focus on the modules instead.

--
1. 44e16b21e852cb88f10798a25956a440a047f7f2
