+++ config/c8/e2e-win.json
@@ -5,8 +5,8 @@
     "**/*.{cjs,d.cts,d.ts}",
     "src/modules/stateless.js",
     "src/modules/testing.js",
-    "src/unix/",
-    "src/unix.js"
+    "src/internal/unix/",
+    "src/internal/unix.js"
   ],
   "check-coverage": false,
   "reports-dir": "_reports/coverage/e2e",
