+++ config/c8/integration-unix.json
@@ -1,7 +1,11 @@
 {
   "all": true,
   "include": ["src/"],
-  "exclude": ["**/*.{cjs,d.cts,d.ts}", "src/win/", "src/win.js"],
+  "exclude": [
+    "**/*.{cjs,d.cts,d.ts}",
+    "src/internal/win/",
+    "src/internal/win.js"
+  ],
   "check-coverage": false,
   "reports-dir": "_reports/coverage/integration",
   "reporter": ["lcov", "text"]
