+++ config/eslint.yml
@@ -545,9 +545,6 @@ overrides:
   # Source code
   - files:
       - src/**/*.js
-      - index.js
-      - stateless.js
-      - testing.js
     rules:
       jsdoc/check-values:
         - error
@@ -774,9 +771,5 @@ ignorePatterns:
   - .temp/
   - node_modules/
   - script/maybe-run.js
-  - src/modules/index.cjs
-  - src/modules/index.d.cts
-  - src/modules/stateless.cjs
-  - src/modules/stateless.d.cts
-  - src/modules/testing.cjs
-  - src/modules/testing.d.cts
+  - src/modules/*.cjs
+  - src/modules/*.d.cts
