+++ config/stryker/integration.js
@@ -3,7 +3,7 @@
 export default {
   coverageAnalysis: "perTest",
   inPlace: false,
-  mutate: ["src/modules/*.js"],
+  mutate: ["src/modules/**/*.js"],
   testRunner: "tap",
   tap: {
     testFiles: ["test/integration/**/*.test.js"],
