+++ config/stryker/unit.js
@@ -3,8 +3,7 @@
 export default {
   coverageAnalysis: "perTest",
   inPlace: false,
-  mutate: ["src/**/*.js"],
-  ignorePatterns: ["src/modules/**/*"],
+  mutate: ["src/internal/**/*.js"],
   testRunner: "tap",
   tap: {
     testFiles: ["test/unit/**/*.test.js"],
