+++ src/internal/executables.js
