+++ src/internal/options.js
