+++ src/internal/platforms.js
