+++ src/internal/reflection.js
