+++ src/internal/unix/bash.js
