+++ src/internal/unix/dash.js
