+++ src/internal/unix/no-shell.js
