+++ src/internal/unix/zsh.js
