+++ src/internal/win.js
