+++ src/internal/win/no-shell.js
