+++ src/internal/win/powershell.js
