+++ test/unit/reflection/is-string.test.js
@@ -6,7 +6,7 @@
 import { testProp } from "@fast-check/ava";
 import * as fc from "fast-check";
 
-import { isString } from "../../../src/reflection.js";
+import { isString } from "../../../src/internal/reflection.js";
 
 testProp("strings", [fc.string()], (t, value) => {
   const result = isString(value);
